QF-211-End to End fixes

+++ src/SFA.DAS.ApprenticeFeedback.Domain/Models/FeedbackTransactionToEmail.cs
@@ -4,7 +4,7 @@ namespace SFA.DAS.ApprenticeFeedback.Domain.Models
 {
     public class FeedbackTransactionToEmail
     {
-        public int ApprenticeFeedbackTransactionId { get; set; }
+        public long ApprenticeFeedbackTransactionId { get; set; }
         public Guid ApprenticeId { get; set; }
     }
 }
